gpu: Take color state into account for glyph lookup
Add a 'linear' bit to the glyph cache key, so we can get have
separate cached images for the two cases. Note that the image
tagging with color state is ignored here - we simply copy the
glyph image out of the atlas as-is, and rely on the cache to
give use the right color state.