gpu: Add a color convert shader
This shader converts between named color states, by using the
same functions that we use on the cpu. The conversion to perform
is passed as an integer encoding the pair of color states. If
we don't have a direction function for going from the first to
the second, we go via XYZ.