cairo: Convert the contents back to srgb
We let cairo operate in linear srgb (by giving it all colors and
texture input in linear srgb), and so we need to convert the
results back to srgb.

This should be rewritten to use color state conversion methods
for GdkTexture when they exist (or at least gdk_memory_convert).