wayland: Add API to restore GdkToplevels
Any toplevel may be referred by an ID, that will be used by the compositor
to restore window state for it. Add the glue so that the upper layers
may restore a toplevel from one of these IDs.

We keep the api private, for now at least.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8502>