demo: Make gtk4-demo use session management
Use it as an example of what is to be expected to make your
application windows restore-able between executions, the session
ID is preserved in a setting between runs, so that it's fed to
the GtkApplication on startup. The actual session ID used is likewise
always stored in this setting, for later runs.

The main window gets an identifier within the session, so that its
state may be restored when first mapped.