gdk/wayland: ignore leave event unless it is for grab
This is another way to work around GNOME/mutter#3546 which is inconsistent
with event delivery from Weston. Given that other compositors may mess this
up perhaps being more defensive on the GDK side is prudent.

Essentially we just ignore a leave event if it has come in for a surface
and we have an explicit grab on another surface. That can happen if we get
a LEAVE event right before getting an ENTER for the grab surface.

Fixes: #6622