snapshot: Don't try to ref null pointer
Was hitting a crash while trying to implement a widget's snapshot with a
`gtk_snapshot_push_mask()`, and found a related issue in the issue
tracker.

It looks like we're okay with the `mask_child` variable being null here
since we do a null-check, but we unconditionally ref it earlier, which
can lead to a crash. Make that ref conditional to fix the crash.

Fixes: https://gitlab.gnome.org/GNOME/gtk/-/issues/6761