[WIP] Jiggle free scrolling for fractional scales
When GTK is rendering at a fractional scale, as we scroll
in integer application pixels, scrolled content lines up
differently with the device grid, and renders differently.
In particular, the snapping that the GL renderer does to
the device grid causes lines to jiggle up and down as
you scroll.

Add gtk_snapshot_push_scroll_offset() which offsets children
in a way that animates nicely by rounding the offset to the
device grid, and use that to implement jiggle-free scrolling
for GtkTextView and GtkViewport. This also provides smooth
device pixel scrolling when the window has integer scale
factor like 200%.