gpu: Add a helper to adjust color states
Add a function that takes an image and a colorstate, and adjusts
the colorstate according to the image flags. This is meant to handle
cases where the upload changes the effective color state of the data,
such as GL_SRGB or importing yuv as external texture.

Note: this function returns a reference.