Add an avif loder
Add a loader for avif images. It uses libavif and produces dmabufs
when /dev/udmabuf is available. Otherwise it produces YUVA memory
textures.

For saving, currently only YUV444 is supported. It would be nice
to just save dmabufs as-is, but that requires more work.

This is meant to facilitate testing of our yuv support, and
therefore, it is not exposed via the public texture loading and
saving apis. It is disabled by default and should *not* be enabled
in distro builds.

avif: Make it build outside of libgtk

We want to use this code in the image tool, and we don't want to
link it statically.

avif loader fixes