css: Add gtk_css_color_value_get_color
This returns a pointer to a GdkColor that is stored inside
the css value. This will simply the callers. Currently the
css color value stores both a GdkRGBA and a GdkColor struct.
This is temporary. The goal is to replace all uses of the rgba
with the color, and then drop the rgba field.