gsk: Track whether nodes require wide gamut
Add gdk_color_state_is_wide_gamut and gsk_render_node_is_hdr, which
tracks whether a render node's content is in a wide gamut color
state (in practice, that means non-sRGB).

This will be used in render_texture to determine the color
state to use when creating a texture.