surface: Add a hdr metadata field
Allow surfaces to carry HDR metadata for describing the
capabilities of the output it is presented on.

This information is important for gamut or tone mapping,
since the goal is to produce colors that can actually be
presented on the output.

When used in this way, the data is sometimes called the
'target color volume'.