Allow defining color states from primaries
The math for deriving the matrix from the primaries is taken
from

https://registry.khronos.org/DataFormat/specs/1.3/dataformat.1.3.html#PRIMARY_CONVERSION

For a colorstate that was constructed from primaries, we return
a cicp tuple of 2/tf/0/1. We interpret the 'unspecified' value
for the primaries to mean: primaries were manually specified.
Returning a non-null cicp tuple is useful, since we have no other
way to communicate the tf.